hv: code style fix for partition mode specific code
There are still some misra-c violations when CONFIG_PARTITION_MODE is defined.

arch/x86/configs:
- remove the unused macro: PRE_LAUNCH_VM_NUM
- pt_dev.c: VMx_CONFIG_PCI_PTDEV_NUM has been defined in partition_config.h,
  should not hard code them again in pt_dev.c.
- ve820.c: use "UL" suffix instead of "U" for 64 bits variables.

vmid is uint16_t in ACRN, so vuart_vmid should be uint16_t as well.

Fix another few other miscellaneous misra-c violations.

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/include/debug/vuart.h
@@ -67,7 +67,7 @@ struct acrn_vuart {
 };
 
 #ifdef CONFIG_PARTITION_MODE
-extern int8_t vuart_vmid;
+extern uint16_t vuart_vmid;
 #endif /* CONFIG_PARTITION_MODE */
 
 struct acrn_vuart *vm_vuart(struct acrn_vm *vm);
